Trying to get tests working.
