--- src/CampaignKit.WorldMap.Core/Class1.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace CampaignKit.WorldMap.Core
-{
-    public class Class1
-    {
-    }
-}
