+++ src/CampaignKit.WorldMap.Core/Services/DefaultProgressService.cs
@@ -21,7 +21,7 @@ using System.Threading.Tasks;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 
-namespace CampaignKit.WorldMap.Services
+namespace CampaignKit.WorldMap.Core.Services
 {
     /// <inheritdoc />
     /// <summary>
