+++ src/CampaignKit.WorldMap.Core/Services/DefaultUserManagerService.cs
@@ -14,7 +14,7 @@
 // limitations under the License.
 // </copyright>
 
-namespace CampaignKit.WorldMap.Services
+namespace CampaignKit.WorldMap.Core.Services
 {
     using System;
     using System.Linq;
