+++ src/CampaignKit.WorldMap.Function/TileCreationTrigger.cs
@@ -1,6 +1,6 @@
 using System;
 
-using CampaignKit.WorldMap.Core;
+using CampaignKit.WorldMap.Core.Services;
 
 using Microsoft.Azure.WebJobs;
 using Microsoft.Extensions.Configuration;
