+++ src/CampaignKit.WorldMap.UI/ViewHelpers/HtmlHelperExtensions.cs
@@ -20,7 +20,7 @@ using System.Linq.Expressions;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.AspNetCore.Mvc.ViewFeatures;
 
-namespace CampaignKit.WorldMap.ViewHelpers
+namespace CampaignKit.WorldMap.UI.ViewHelpers
 {
     /// <summary>
     ///     Class HtmlHelperExtensions.
