+++ src/CampaignKit.WorldMap.UI/ViewModels/MapEditViewModel.cs
@@ -14,7 +14,7 @@
 // limitations under the License.
 // </copyright>
 
-namespace CampaignKit.WorldMap.ViewModels
+namespace CampaignKit.WorldMap.UI.ViewModels
 {
     using System.ComponentModel.DataAnnotations;
     using Microsoft.AspNetCore.Mvc;
