+++ src/CampaignKit.WorldMap.UI/ViewModels/MarkerEditViewModel.cs
@@ -14,7 +14,7 @@
 // limitations under the License.
 // </copyright>
 
-namespace CampaignKit.WorldMap.ViewModels
+namespace CampaignKit.WorldMap.UI.ViewModels
 {
     /// <summary>
     ///     Class MapShowViewModel.
