+++ src/CampaignKit.WorldMap.UI/Views/_ViewImports.cshtml
@@ -13,8 +13,8 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 -->
-@using CampaignKit.WorldMap.Entities
-@using CampaignKit.WorldMap.ViewHelpers
-@using CampaignKit.WorldMap.ViewModels
+@using CampaignKit.WorldMap.Core.Entities
+@using CampaignKit.WorldMap.UI.ViewHelpers
+@using CampaignKit.WorldMap.UI.ViewModels
 
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
