+++ tests/CampaignKit.WorldMap.Tests/CampaignKit.WorldMap.Tests.csproj
@@ -46,8 +46,8 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\CampaignKit.WorldMap.Core\CampaignKit.WorldMap.Core.csproj" />
-    <ProjectReference Include="..\..\src\CampaignKit.WorldMap.TileProcessor\CampaignKit.WorldMap.Function.csproj" />
-    <ProjectReference Include="..\..\src\CampaignKit.WorldMap\CampaignKit.WorldMap.UI.csproj" />
+    <ProjectReference Include="..\..\src\CampaignKit.WorldMap.Function\CampaignKit.WorldMap.Function.csproj" />
+    <ProjectReference Include="..\..\src\CampaignKit.WorldMap.UI\CampaignKit.WorldMap.UI.csproj" />
   </ItemGroup>
 
 </Project>
