+++ tests/CampaignKit.WorldMap.Tests/MockServices/MockBlobStorageService.cs
@@ -1,9 +1,6 @@
-﻿using CampaignKit.WorldMap.Core;
-using CampaignKit.WorldMap.Services;
+﻿using CampaignKit.WorldMap.Core.Services;
 
 using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 
 namespace CampaignKit.WorldMap.Tests.MockServices
